refactor: streamline license terms registration by implementing setMaxLicenseTokens utility and updating related methods for improved clarity and consistency
