+++ packages/core-sdk/test/unit/resources/ipAsset.test.ts
@@ -1747,177 +1747,6 @@ describe("Test IpAssetClient", () => {
     });
   });
 
-  describe("Test ipAssetClient.registerPilTermsAndAttach", () => {
-    it("should throw ipId error when registerPilTermsAndAttach given ipId is wrong address", async () => {
-      try {
-        await ipAssetClient.registerPilTermsAndAttach({
-          ipId: "0x",
-          licenseTermsData: [
-            {
-              terms: licenseTerms,
-              licensingConfig,
-            },
-          ],
-        });
-      } catch (err) {
-        expect((err as Error).message).equal(
-          "Failed to register PIL terms and attach: Invalid address: 0x.",
-        );
-      }
-    });
-
-    it("should throw ipId have not registered error when registerPilTermsAndAttach given ipId have not registered", async () => {
-      stub(ipAssetClient.ipAssetRegistryClient, "isRegistered").resolves(false);
-
-      try {
-        await ipAssetClient.registerPilTermsAndAttach({
-          ipId: "0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c",
-          licenseTermsData: [
-            {
-              terms: licenseTerms,
-              licensingConfig,
-            },
-          ],
-        });
-      } catch (err) {
-        expect((err as Error).message).equal(
-          "Failed to register PIL terms and attach: The IP with id 0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c is not registered.",
-        );
-      }
-    });
-
-    it("should return encoded tx data when registerPilTermsAndAttach given correct args and encodedTxDataOnly of true", async () => {
-      stub(ipAssetClient.ipAssetRegistryClient, "isRegistered").resolves(true);
-
-      stub(ipAssetClient.licenseTemplateClient, "getLicenseTermsId").resolves({
-        selectedLicenseTermsId: 0n,
-      });
-
-      const result = await ipAssetClient.registerPilTermsAndAttach({
-        ipId: "0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c",
-        licenseTermsData: [
-          {
-            terms: licenseTerms,
-            licensingConfig,
-          },
-        ],
-        txOptions: {
-          encodedTxDataOnly: true,
-        },
-      });
-
-      expect(result.encodedTxData!.data).to.be.a("string");
-    });
-
-    it("should return txHash when registerPilTermsAndAttach given correct args ", async () => {
-      stub(ipAssetClient.licenseAttachmentWorkflowsClient, "registerPilTermsAndAttach").resolves(
-        txHash,
-      );
-      stub(ipAssetClient.ipAssetRegistryClient, "isRegistered").resolves(true);
-
-      stub(ipAssetClient.licenseTemplateClient, "getLicenseTermsId").resolves({
-        selectedLicenseTermsId: 0n,
-      });
-      const result = await ipAssetClient.registerPilTermsAndAttach({
-        ipId: "0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c",
-        licenseTermsData: [
-          {
-            terms: licenseTerms,
-            licensingConfig,
-          },
-          {
-            terms: licenseTerms,
-            licensingConfig,
-          },
-        ],
-      });
-
-      expect(result.txHash).to.equal(txHash);
-      expect(result.licenseTermsIds).to.deep.equal([0n, 0n]);
-      expect(result.maxLicenseTokensTxHashes).to.be.an("undefined");
-    });
-
-    it("should return licenseTermsMaxLimitTxHashes when registerPilTermsAndAttach given correct args with license terms max limit", async () => {
-      stub(ipAssetClient.licenseAttachmentWorkflowsClient, "registerPilTermsAndAttach").resolves(
-        txHash,
-      );
-      stub(ipAssetClient.ipAssetRegistryClient, "isRegistered").resolves(true);
-      stub(ipAssetClient.licenseTemplateClient, "getLicenseTermsId").resolves({
-        selectedLicenseTermsId: 1n,
-      });
-      const result = await ipAssetClient.registerPilTermsAndAttach({
-        ipId: "0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c",
-        licenseTermsData: [
-          {
-            terms: licenseTerms,
-            licensingConfig,
-            maxLicenseTokens: 100,
-          },
-          {
-            terms: licenseTerms,
-            licensingConfig,
-            maxLicenseTokens: 100,
-          },
-          {
-            terms: licenseTerms,
-            licensingConfig,
-          },
-        ],
-      });
-      expect(result.maxLicenseTokensTxHashes).to.be.an("array");
-      expect(result.maxLicenseTokensTxHashes?.length).to.be.equal(2);
-    });
-
-    it("should call with default values of licensingConfig when registerPilTermsAndAttach given licensingConfig is not provided", async () => {
-      const registerPilTermsAndAttachStub = stub(
-        ipAssetClient.licenseAttachmentWorkflowsClient,
-        "registerPilTermsAndAttach",
-      ).resolves(txHash);
-      stub(ipAssetClient.ipAssetRegistryClient, "isRegistered").resolves(true);
-
-      await ipAssetClient.registerPilTermsAndAttach({
-        ipId: "0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c",
-        licenseTermsData: [
-          {
-            terms: licenseTerms,
-          },
-        ],
-      });
-      expect(
-        registerPilTermsAndAttachStub.args[0][0].licenseTermsData[0].licensingConfig,
-      ).to.deep.equal({
-        isSet: false,
-        mintingFee: 0n,
-        licensingHook: zeroAddress,
-        hookData: zeroHash,
-        commercialRevShare: 0,
-        disabled: false,
-        expectMinimumGroupRewardShare: 0,
-        expectGroupRewardPool: zeroAddress,
-      });
-    });
-
-    it("should be called with expected values given PILFlavor.nonCommercialSocialRemixing", async () => {
-      const registerPilTermsAndAttachStub = stub(
-        ipAssetClient.licenseAttachmentWorkflowsClient,
-        "registerPilTermsAndAttach",
-      ).resolves(txHash);
-      stub(ipAssetClient.ipAssetRegistryClient, "isRegistered").resolves(true);
-
-      await ipAssetClient.registerPilTermsAndAttach({
-        ipId: "0x1daAE3197Bc469Cb97B917aa460a12dD95c6627c",
-        licenseTermsData: [
-          {
-            terms: PILFlavor.nonCommercialSocialRemixing(),
-          },
-        ],
-      });
-      expect(registerPilTermsAndAttachStub.args[0][0].licenseTermsData[0].terms).to.deep.equal(
-        PILFlavor.nonCommercialSocialRemixing(),
-      );
-    });
-  });
-
   describe("Test ipAssetClient.mintAndRegisterIpAndMakeDerivativeWithLicenseTokens", () => {
     it("should throw licenseTokens error when mintAndRegisterIpAndMakeDerivativeWithLicenseTokens given licenseTokens empty", async () => {
       try {
